+++ src/binary/init/chess_module.rs
