+++ src/binary/init/console_module.rs
