+++ src/binary/init/fmt_module.rs
