+++ src/binary/init/fn_module.rs
