+++ src/binary/init/pipe_module.rs
