+++ src/binary/init/shell_module.rs
