+++ src/binary/init/widget_module.rs
